refactor code modules
