+++ src/lib/settings/index.ts
@@ -1,2 +1,3 @@
+export * from './manager'
 export * from './settings'
 export * from './static'
