+++ src/lib/settings/settings.ts
@@ -4,116 +4,10 @@ import * as Lo from 'lodash'
 import { Primitive } from 'type-fest'
 import { inspect } from 'util'
 import { IsRecord, PlainObject } from '../utils'
-import { DataDefault, Fixup, MapType, Shorthand, Spec, Validate } from './static'
+import type { Options } from './manager'
+import { Fixup, MapType, Shorthand, Validate } from './static'
 
 const log = Logger.log.child('settings')
-
-/**
- * API
- */
-
-/**
- * todo
- */
-export type Manager<Input extends PlainObject, Data extends PlainObject> = {
-  reset(): Manager<Input, Data>
-  change(input: Input): Manager<Input, Data>
-  original(): Data
-  metadata: MetadataState<Data>
-  data: Data
-}
-
-// todo errors currently report names of setting fields, but not the namespaces
-// to it (if any)
-// todo should onFixup be replaced with a batch version of onfixups that gets
-// called with all fixups that happened for all of the input?
-// todo allow env vars to populate settings
-// todo track env var as value source
-// todo $initial magic var to reset settting to its original state, re-running
-// dynamic initializers if necessary
-// todo run initial through fixup in dev to be safer
-// todo run initial through validation in dev to be safer
-
-export type FixupInfo = {
-  name: string
-  before: unknown
-  after: unknown
-  messages: string[]
-}
-
-export type Options = {
-  /**
-   * Handle fixup events.
-   *
-   * If your settings spec has no fixups then you can ignore this option.
-   *
-   * By default, fixups are logged at warning level. If you provide your own
-   * function then this default behaviour will be disabled. You can retain it by
-   * calling the default function passed as a second argument to your function.
-   */
-  onFixup?: (info: FixupInfo, originalHandler: (info: FixupInfo) => void) => void
-}
-
-/**
- * Default onFixup handler.
- */
-function onFixup(info: FixupInfo): void {
-  log.warn(
-    'One of your setting values was invalid. We were able to automaticlaly fix it up now but please update your code.',
-    info
-  )
-}
-
-/**
- *
- */
-export function create<Input extends PlainObject, Data extends PlainObject = DataDefault<Input>>({
-  fields,
-  ...options
-}: {
-  fields: Spec<Input, Data>
-} & Options): Manager<Input, Data> {
-  log.debug('construct')
-  if (isDevelopment()) {
-    validateSpecifier({ fields }, { path: '__root__' })
-  }
-
-  // todo we currently have to clone the given spec deeply because mapEntryData mutations the spec with shadow specifiers
-  // and shodow specifiers currently break the second+ initialize run (e.g. during reset)
-  // todo we didn't catch this yet with own unit test, but other nexus unit tets caught the issue
-
-  const initial = initialize({ fields: Lo.cloneDeep(fields) }, { path: '__root__' })
-  const state = {
-    data: initial.data as Data,
-    original: (undefined as any) as Data, // lazy
-    metadata: initial.metadata as any, // Metadata<Data>,
-  }
-
-  const api: Manager<Input, Data> = {
-    data: state.data,
-    metadata: state.metadata,
-    change(input) {
-      log.debug('change', { input })
-      const newData = resolve(options, 'set', { fields }, input, state.data, state.metadata)
-      commit({ fields: Lo.cloneDeep(fields) }, 'set', newData, state.data, state.metadata)
-      return api
-    },
-    reset() {
-      log.debug('reset')
-      const initial = initialize({ fields: Lo.cloneDeep(fields) }, { path: '__root__' })
-      api.data = state.data = initial.data as any
-      api.metadata = state.metadata = initial.metadata as any
-      return api
-    },
-    original() {
-      log.debug('get original')
-      const original = state.original ?? dataFromMetadara(state.metadata, {})
-      return original
-    },
-  }
-
-  return api
-}
 
 /**
  * Metadata
@@ -293,6 +187,23 @@ function resolveRecord(
   return newData
 }
 
+export type FixupInfo = {
+  name: string
+  before: unknown
+  after: unknown
+  messages: string[]
+}
+
+/**
+ * Default onFixup handler.
+ */
+function onFixup(info: FixupInfo): void {
+  log.warn(
+    'One of your setting values was invalid. We were able to automaticlaly fix it up now but please update your code.',
+    info
+  )
+}
+
 function resolveLeaf(options: Options, specifier: SpecifierLeaf, input: any, info: TraversalInfo): any {
   let resolvedValue = input
 
@@ -370,7 +281,7 @@ function resolveLeaf(options: Options, specifier: SpecifierLeaf, input: any, inf
  * fixups, validation and so on until finally assigning it into the setting data.
  * The input is not mutated. The data is.
  */
-function resolve(
+export function resolve(
   options: Options,
   metadataFrom: MetadataValueFromType,
   parentSpecifier: SpecifierNamespace,
@@ -447,7 +358,7 @@ function resolve(
 /**
  *
  */
-function commit(
+export function commit(
   specifier: SpecifierNamespace,
   metadataFrom: MetadataValueFromType,
   input: any,
@@ -561,16 +472,16 @@ function doCommitRecord(
  * specifiers
  */
 
-type Specifier = SpecifierLeaf | SpecifierRecord | SpecifierNamespace
+export type Specifier = SpecifierLeaf | SpecifierRecord | SpecifierNamespace
 
 /**
  *
  */
-function isLeafSpecifier(specifier: any): specifier is SpecifierLeaf {
+export function isLeafSpecifier(specifier: any): specifier is SpecifierLeaf {
   return !isNamespaceSpecifier(specifier) && !isRecordSpecifier(specifier)
 }
 
-type SpecifierLeaf = {
+export type SpecifierLeaf = {
   mapType?: MapType
   fixup?: Fixup
   validate?: Validate
@@ -579,11 +490,11 @@ type SpecifierLeaf = {
 /**
  *
  */
-function isRecordSpecifier(specifier: any): specifier is SpecifierRecord {
+export function isRecordSpecifier(specifier: any): specifier is SpecifierRecord {
   return Boolean(specifier?.entry)
 }
 
-type SpecifierRecord = {
+export type SpecifierRecord = {
   entry: Specifier
   mapEntryData?(newEntryData: any, entryKey: string): any
 }
@@ -591,11 +502,11 @@ type SpecifierRecord = {
 /**
  *
  */
-function isNamespaceSpecifier(specifier: any): specifier is SpecifierNamespace {
+export function isNamespaceSpecifier(specifier: any): specifier is SpecifierNamespace {
   return Boolean(specifier?.fields)
 }
 
-type SpecifierNamespace = {
+export type SpecifierNamespace = {
   fields: any
   initial?(): any
   shorthand?: Shorthand
@@ -610,7 +521,7 @@ type InitializeResult = { data: PlainObject; metadata: Metadata }
 /**
  *
  */
-function initialize(specifier: Specifier, info: TraversalInfo): InitializeResult {
+export function initialize(specifier: Specifier, info: TraversalInfo): InitializeResult {
   if (isNamespaceSpecifier(specifier)) return initializeNamespace(specifier, info)
   if (isRecordSpecifier(specifier)) return initializeRecord(specifier, info)
   if (isLeafSpecifier(specifier)) return initializeLeaf(specifier, info)
@@ -828,32 +739,6 @@ function runInitializer(specifier: any, info: TraversalInfo): any {
 }
 
 /**
- * Validate the spec for basic invariants.
- * todo mapData mapEntryData fixup etc.
- */
-function validateSpecifier(specifier: Specifier, info: TraversalInfo) {
-  if (isNamespaceSpecifier(specifier)) {
-    Lo.forOwn(specifier.fields, (fieldSpecifier: Specifier, fieldName: string) => {
-      validateSpecifier(fieldSpecifier, { path: fieldName })
-    })
-    return
-  }
-
-  if (isRecordSpecifier(specifier)) {
-    validateSpecifier(specifier.entry as SpecifierNamespace, info)
-    return
-  }
-
-  if (specifier.mapType !== undefined && typeof specifier.mapType !== 'function') {
-    throw new Error(
-      `Type mapper for setting "${info.path}" was invalid. Type mappers must be functions. Got: ${inspect(
-        specifier.mapType
-      )}`
-    )
-  }
-}
-
-/**
  * utils
  */
 
@@ -861,22 +746,6 @@ function validateSpecifier(specifier: Specifier, info: TraversalInfo) {
 // function renderPath(info:TraversalInfo): string {
 //   return info.path.slice(1).join('.')
 // }
-
-/**
- * Check if curerntly in production mode defined as
- * NODE_ENV environment variable equaling "production".
- */
-function isProduction(): boolean {
-  return process.env.NODE_ENV === 'production'
-}
-
-/**
- * Check if curerntly in development mode defined as
- * NODE_ENV environment variable not equaling "production".
- */
-function isDevelopment(): boolean {
-  return process.env.NODE_ENV !== 'production'
-}
 
 function mergeShallow(o1: any, o2: any) {
   for (const [k, v] of Object.entries(o2)) {
@@ -890,10 +759,10 @@ function mergeShallow(o1: any, o2: any) {
 /**
  *
  */
-function dataFromMetadara<Data>(metadata: MetadataNamespace, copy: PlainObject): Data {
+export function dataFromMetadata<Data>(metadata: MetadataNamespace, copy: PlainObject): Data {
   Lo.forOwn(metadata.fields, (fieldMetadata, name) => {
     if ('fields' in fieldMetadata) {
-      copy[name] = dataFromMetadara(fieldMetadata, {})
+      copy[name] = dataFromMetadata(fieldMetadata, {})
     } else {
       copy[name] = fieldMetadata.initial
     }
@@ -902,6 +771,6 @@ function dataFromMetadara<Data>(metadata: MetadataNamespace, copy: PlainObject):
   return copy as any
 }
 
-type TraversalInfo = {
+export type TraversalInfo = {
   path: string
 }
